+++ CHANGELOG.md
@@ -60,6 +60,11 @@ Bug Fixes
   not quite, flat (#741)
 - Fixed a bug in chromatic parametric galaxy models based on COSMOS galaxies.
   (#745)
+- Fixed a bug in the Image copy constructor where the wcs was not copied if
+  the `image` parameter is a galsim.Image. (#762)
+- Fixed a bug in the Sum and Convolution constructors when they are only
+  adding or convolving a single element that could lead to erroneous str and
+  reprs for the resulting object if it was then transformed. (#763)
 
 
 Deprecated Features